fix(gates): split barrier runtime-store helpers for the file cap
